Add bones of DVCReader interface
This is the interface between Python DVC and this Node extension.
Also has the start of a function to infer default config from a
project directory.

At least for now, it works by just running DVC on the CLI and parsing
JSON output.
Also includes setup for jest-ts unit tests.