Make ExperimentsTable refresh return a void promise
This scales better with additional ways to read the experiments data, and the
one place we use `refresh` uses it as a void promise anyway.

The underlying private implementation is also tweaked, the functionality is the
same but now the promises returned by refresh should behave more sanely (one
promise = one update) and we can add parsing logic to `updateData`.

Existing names are kept to simplify the diff, we can rename `refresh` in a
follow-up if we want.