Remove demo project target for dev UI scripts
Without a provided project, the dev UI loads up the last open project, or a
blank window otherwise. This can greatly benefit the workflow of testing the
extension against projects that aren't the built-in demo, and also paves the way
for removing the embedded demo repo if we decide that's best.